Hack a fix to blockhash off by one error
When you create an ephemeral state, calls to `blockhash()` are off by one.
This is pretty bad and I'm not sure how to fix it. Instead of solving
the issue at its root, I no longer use `get_recommended_target_hash()`
and instead calculate the hash on my own. This isn't a good permanent
solution but it means I can move on and work on some other stuff for a
bit.

This also adds an extra alloc account for validation. Once again
hardcoded and should be replaced by a genesis declaration.